+++ lib/rohd_hcl.dart
@@ -2,13 +2,11 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 export 'src/adder.dart';
-export 'src/sum.dart';
 export 'src/arbiters/arbiters.dart';
 export 'src/binary_gray.dart';
 export 'src/carry_save_mutiplier.dart';
 export 'src/component_config/component_config.dart';
 export 'src/count.dart';
-export 'src/counter.dart';
 export 'src/edge_detector.dart';
 export 'src/encodings/encodings.dart';
 export 'src/error_checking/error_checking.dart';
@@ -24,4 +22,5 @@ export 'src/ripple_carry_adder.dart';
 export 'src/rotate.dart';
 export 'src/shift_register.dart';
 export 'src/sort.dart';
+export 'src/summation/summation.dart';
 export 'src/utils.dart';
