+++ lib/src/summation/sum.dart
@@ -7,70 +7,12 @@
 // 2024 August 26
 // Author: Max Korbel <max.korbel@intel.com>
 
-import 'dart:math';
-
 import 'package:collection/collection.dart';
 import 'package:rohd/rohd.dart';
 import 'package:rohd_hcl/rohd_hcl.dart';
-import 'package:rohd_hcl/src/exceptions.dart';
-import 'package:rohd_hcl/src/parallel_prefix_operations.dart';
+import 'package:rohd_hcl/src/summation/summation_utils.dart';
 
-class SumInterface extends PairInterface {
-  final bool hasEnable;
-
-  /// The [amount] to increment/decrement by, depending on [increments].
-  late final Logic amount =
-      fixedAmount != null ? Const(fixedAmount, width: width) : port('amount');
-
-  /// Controls whether it should increment or decrement (based on [increments])
-  /// (active high).
-  ///
-  /// Present if [hasEnable] is `true`.
-  Logic? get enable => tryPort('enable');
-
-  final int width;
-
-  /// If `true`, will increment. If `false`,  will decrement.
-  final bool increments;
-
-  final dynamic fixedAmount;
-
-  /// TODO
-  ///
-  /// If [width] is `null`, it can be inferred from [fixedAmount] if provided
-  /// with a type that contains width information (e.g. a [LogicValue]). There
-  /// must be enough information provided to determine the [width].
-  ///
-  /// If a [fixedAmount] is provided, then [amount] will be tied to a [Const]. A
-  /// provided [fixedAmount] must be parseable by [LogicValue.of]. Note that the
-  /// [fixedAmount] will always be interpreted as a positive value truncated to
-  /// [width]. If no [fixedAmount] is provided, then [amount] will be a normal
-  /// [port] with [width] bits.
-  ///
-  /// If [hasEnable] is `true`, then an [enable] port will be added to the
-  /// interface.
-  SumInterface(
-      {this.fixedAmount,
-      this.increments = true,
-      int? width,
-      this.hasEnable = false})
-      : width = width ?? LogicValue.ofInferWidth(fixedAmount).width {
-    setPorts([
-      if (fixedAmount == null) Port('amount', this.width),
-      if (hasEnable) Port('enable'),
-    ], [
-      PairDirection.fromProvider
-    ]);
-  }
-
-  SumInterface.clone(SumInterface other)
-      : this(
-          fixedAmount: other.fixedAmount,
-          increments: other.increments,
-          width: other.width,
-          hasEnable: other.hasEnable,
-        );
-
+extension on SumInterface {
   List<Conditional> _combAdjustments(Logic Function(Logic) s, Logic nextVal) {
     final conds = <Conditional>[
       if (increments)
@@ -87,25 +29,25 @@ class SumInterface extends PairInterface {
   }
 }
 
-class Sum extends Module with DynamicInputToLogic {
+class Sum extends Module with DynamicInputToLogicForSummation {
   final int width;
 
   /// If `true`,  will saturate at the `maxValue` and `minValue`. If `false`,
   /// will wrap around (overflow/underflow) at the `maxValue` and `minValue`.
   final bool saturates;
 
-  Logic get value => output('value');
+  Logic get sum => output('sum');
 
   /// Indicates whether the sum has reached the maximum value.
   ///
-  /// If it [saturates], then [value] will be equal to the maximum value.
+  /// If it [saturates], then [sum] will be equal to the maximum value.
   /// Otherwise, the value may have overflowed to any value, but the net sum
   /// before overflow will have been greater than the maximum value.
   Logic get reachedMax => output('reachedMax');
 
   /// Indicates whether the sum has reached the minimum value.
   ///
-  /// If it [saturates], then [value] will be equal to the minimum value.
+  /// If it [saturates], then [sum] will be equal to the minimum value.
   /// Otherwise, the value may have underflowed to any value, but the net sum
   /// before underflow will have been less than the minimum value.
   Logic get reachedMin => output('reachedMin');
@@ -166,7 +108,7 @@ class Sum extends Module with DynamicInputToLogic {
               uniquify: (original) => '${original}_$i'))
         .toList();
 
-    addOutput('value', width: this.width);
+    addOutput('sum', width: this.width);
     addOutput('reachedMax');
     addOutput('reachedMin');
 
@@ -216,7 +158,7 @@ class Sum extends Module with DynamicInputToLogic {
       ..gets(minValueLogic + zeroPoint);
 
     final internalValue = Logic(name: 'internalValue', width: internalWidth);
-    value <= (internalValue - zeroPoint).getRange(0, this.width);
+    sum <= (internalValue - zeroPoint).getRange(0, this.width);
 
     final passedMax = Logic(name: 'passedMax');
     final passedMin = Logic(name: 'passedMin');
@@ -265,50 +207,4 @@ class Sum extends Module with DynamicInputToLogic {
   }
 
   static int _biggestVal(int width) => (1 << width) - 1;
-}
-
-//TODO doc
-//TODO: hide this somehow
-int inferWidth(
-    List<dynamic> values, int? width, List<SumInterface> interfaces) {
-  if (width != null) {
-    if (width <= 0) {
-      throw RohdHclException('Width must be greater than 0.');
-    }
-
-    if (values.any((v) => v is Logic && v.width > width)) {
-      throw RohdHclException(
-          'Width must be at least as large as the largest value.');
-    }
-
-    return width;
-  }
-
-  int? maxWidthFound;
-
-  for (final value in values) {
-    int? inferredValWidth;
-    if (value is Logic) {
-      inferredValWidth = value.width;
-    } else if (value != null) {
-      inferredValWidth = LogicValue.ofInferWidth(value).width;
-    }
-
-    if (inferredValWidth != null &&
-        (maxWidthFound == null || inferredValWidth > maxWidthFound)) {
-      maxWidthFound = inferredValWidth;
-    }
-  }
-
-  for (final interface in interfaces) {
-    if (interface.width > maxWidthFound!) {
-      maxWidthFound = interface.width;
-    }
-  }
-
-  if (maxWidthFound == null) {
-    throw RohdHclException('Unabled to infer width.');
-  }
-
-  return max(1, maxWidthFound);
 }
