+++ lib/src/summation/summation.dart
@@ -1,0 +1,6 @@
+// Copyright (C) 2024 Intel Corporation
+// SPDX-License-Identifier: BSD-3-Clause
+
+export 'counter.dart';
+export 'sum.dart';
+export 'sum_interface.dart';
